refactor: Update import paths for verifyToken and remove establishment routes
